% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_theme}
\alias{add_theme}
\title{Add color and plot theme}
\usage{
add_theme(
  plot,
  aesthetic = c("color", "fill"),
  discrete = TRUE,
  palette = c("1", "2"),
  ...
)
}
\arguments{
\item{plot}{A ggplot object}

\item{aesthetic}{One of "color" or "fill" indicating which aesthetic to add
color to}

\item{discrete}{Logical indicating whether to use the discrete or continuous
color/fill scale}

\item{palette}{One of 1 or 2 indicating which default color palette to use}

\item{...}{Additional arguments to pass to the color/fill scale functions}
}
\value{
A ggplot object with color/plot theme added
}
\description{
Add color and plot theme
}
\keyword{internal}
