% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_aesthetics}
\alias{get_aesthetics}
\title{Get aesthetics layer to add to ggplot.}
\usage{
get_aesthetics(...)
}
\arguments{
\item{...}{Named arguments where the name of the argument is formatted as
\{aesthetic_name\}_str and the value is the name of the variable in the data
frame to use for that aesthetic. For example, to specify the x-axis
variable, use \code{x_str = "variable_name"}.}
}
\value{
A \code{\link[ggplot2:aes]{ggplot2::aes()}} object.
}
\description{
Helper function to ignore \code{NULL} inputs when adding
aesthetics to a ggplot.
}
