% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-plots.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{Boxplot wrapper for ggplot2}
\usage{
plot_boxplot(
  data,
  x_str = NULL,
  y_str = NULL,
  fill_str = NULL,
  horizontal = FALSE,
  ...
)
}
\arguments{
\item{data}{Data frame to use for plot.}

\item{x_str}{Character string (optional). Name of variable in \code{data} to
plot on x-axis. Should be discrete (e.g., a factor variable).}

\item{y_str}{Character string. Name of variable in \code{data} to plot on
y-axis. If \code{NULL}, plot boxplot using all values in data frame.}

\item{fill_str}{Character string. Name of variable in \code{data}
to use as fill aesthetic in plot.}

\item{horizontal}{Logical. Whether the boxplots should be horizontal instead
of vertical.}

\item{...}{Other arguments to pass to \code{\link[ggplot2:geom_boxplot]{ggplot2::geom_boxplot()}}.}
}
\value{
A ggplot boxplot.
}
\description{
Convenient wrapper around \code{\link[ggplot2:geom_boxplot]{ggplot2::geom_boxplot()}}.
}
\examples{
## plot boxplot of all data in data frame
plot_boxplot(as.data.frame(matrix(rnorm(1000), nrow = 100)))
## plot boxplot of single column in data frame
plot_boxplot(iris, y_str = "Sepal.Width")
plot_boxplot(iris, x_str = "Species", y_str = "Sepal.Width")
iris2 <- data.frame(
  iris,
  z = as.factor(rep(letters[1:2], length.out = nrow(iris)))
)
plot_boxplot(iris2, x_str = "Species", y_str = "Sepal.Width", fill_str = "z")
plot_boxplot(iris2, y_str = "Sepal.Width", fill_str = "z")

}
