% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-plots.R
\name{plot_density}
\alias{plot_density}
\title{Kernel density plot wrapper for ggplot2}
\usage{
plot_density(
  data,
  x_str = NULL,
  fill_str = NULL,
  fill = "#6FBBE3",
  alpha = 0.4,
  ...
)
}
\arguments{
\item{data}{Data frame to use for plot.}

\item{x_str}{Character string. Name of variable in \code{data} to plot on
x-axis. If \code{NULL}, plot density using all values in data frame.}

\item{fill_str}{Character string. Name of variable in \code{data}
to use as fill aesthetic in plot.}

\item{fill}{Fill color. Only used if \code{fill_str} is \code{NULL}.}

\item{alpha}{Alpha value for transparency.}

\item{...}{Other arguments to pass to \code{\link[ggplot2:geom_density]{ggplot2::geom_density()}}.}
}
\value{
A ggplot density.
}
\description{
Convenient wrapper around \code{\link[ggplot2:geom_density]{ggplot2::geom_density()}}.
}
\examples{
## plot distribution of all data in data frame
plot_density(as.data.frame(rnorm(1000), nrow = 100))
## plot distribution of a single column in data frame
plot_density(iris, x_str = "Sepal.Width")
plot_density(iris, x_str = "Sepal.Width", fill_str = "Species")

}
