% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust-plots.R
\name{plot_hclust}
\alias{plot_hclust}
\title{Plot hierarchical clustering dendrograms}
\usage{
plot_hclust(
  data,
  leaf_labels = rownames(data),
  leaf_colors = NULL,
  dist_metric = "euclidean",
  dist_matrix = NULL,
  linkage = "ward.D",
  text_size = 10,
  title = NULL
)
}
\arguments{
\item{data}{Data matrix or data.frame on which to perform hierarchical
clustering.}

\item{leaf_labels}{(Optional) Text labels for leaf nodes (e.g.,
class/outcome labels).}

\item{leaf_colors}{(Optional) Data vector to use for coloring leaf nodes.}

\item{dist_metric}{Distance metric for clustering (see \code{\link[stats:dist]{stats::dist()}}).}

\item{dist_matrix}{(Optional) Distance matrix for clustering. Must provide
either \code{dist_metric} or \code{dist_matrix}.}

\item{linkage}{Type of linkage for hierarchical clustering (see
\code{\link[stats:hclust]{stats::hclust()}}).}

\item{text_size}{Numeric; size of text for leaf node labels.}

\item{title}{Character string. Title of plot.}
}
\value{
A list of three:
\describe{
\item{plot}{A ggplot object of the dendrogram.}
\item{hclust}{Output of \code{\link[stats:hclust]{stats::hclust()}}}
\item{dend}{Hierarchical clustering dendrogram data. See output of
\code{\link[ggdendro:dendro_data]{ggdendro::dendro_data()}}.}
}
}
\description{
Run hierarchical clustering on the given data matrix and
plot the resulting dendrogram using a pretty theme. Also allows for
coloring of the leaf nodes.
}
\examples{
out <- plot_hclust(data = iris[, -5], leaf_labels = iris$Species,
                  leaf_colors = iris$Species)
out$plot

}
