% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-plots.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Histogram plot wrapper for ggplot2}
\usage{
plot_histogram(
  data,
  x_str = NULL,
  fill_str = NULL,
  fill = "#6FBBE3",
  bins = 12,
  ...
)
}
\arguments{
\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{x_str}{Character string. Name of variable in \code{data} to plot on
x-axis. If \code{NULL}, plot density using all values in data frame.}

\item{fill_str}{Character string. Name of variable in \code{data}
to use as fill aesthetic in plot.}

\item{fill}{Fill color. Only used if \code{fill_str} is \code{NULL}.}

\item{bins}{Number of bins. Overridden by \code{binwidth}. Defaults to 30.}

\item{...}{Other arguments to pass to \code{\link[ggplot2:geom_histogram]{ggplot2::geom_histogram()}}.}
}
\value{
A ggplot histogram.
}
\description{
Convenient wrapper around \code{\link[ggplot2:geom_histogram]{ggplot2::geom_histogram()}}.
}
\examples{
## plot distribution of all data in data frame
plot_histogram(as.data.frame(rnorm(1000), nrow = 100))
## plot distribution of a single column in data frame
plot_histogram(iris, x_str = "Sepal.Width")
plot_histogram(iris, x_str = "Sepal.Width", fill_str = "Species")

}
