% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-plots.R
\name{plot_horizontal_dotplot}
\alias{plot_horizontal_dotplot}
\title{Plot horizontal dot plot.}
\usage{
plot_horizontal_dotplot(
  data,
  x_str,
  y_str,
  color_str = NULL,
  line_size = 1,
  line_color = "darkgray",
  point_size = 8,
  ...
)
}
\arguments{
\item{data}{Data frame to use for plot.}

\item{x_str}{Character string. Name of variable in \code{data} to plot on
x-axis.}

\item{y_str}{Character string. Name of variable in \code{data} to plot on
y-axis.}

\item{color_str}{Character string. Name of variable in \code{data}
to use as color aesthetic in plot.}

\item{line_size}{Size of line, representing range of values.}

\item{line_color}{Color of line, representing range of values.}

\item{point_size}{Size of point.}

\item{...}{Other arguments to pass to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}.}
}
\value{
A ggplot object
}
\description{
Creates a horizontal dot plot, as an alternative to grouped bar
charts.
}
\examples{
df <- data.frame(ID = rep(paste0("ID", 1:3), 2),
                 Group = rep(paste0("Group", LETTERS[1:2]), each = 3),
                 Value = c(1, 2, 3, -1, 0, 4))
plot_horizontal_dotplot(df, x_str = "Value", y_str = "ID", color_str = "Group")

}
