% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-plots.R
\name{plot_line}
\alias{plot_line}
\title{Line plot wrapper for ggplot2}
\usage{
plot_line(data, x_str, y_str, color_str = NULL, linetype_str = NULL, ...)
}
\arguments{
\item{data}{Data frame to use for plot.}

\item{x_str}{Character string. Name of variable in \code{data} to plot on
x-axis.}

\item{y_str}{Character string. Name of variable in \code{data} to plot on
y-axis.}

\item{color_str}{Character string. Name of variable in \code{data}
to use as color aesthetic in plot.}

\item{linetype_str}{Character string. Name of variable in \code{data}
to use as linetype aesthetic in plot.}

\item{...}{Other arguments to pass to \code{\link[ggplot2:geom_path]{ggplot2::geom_line()}}.}
}
\value{
A ggplot line plot.
}
\description{
Convenient wrapper around \code{\link[ggplot2:geom_path]{ggplot2::geom_line()}}.
}
\examples{
df <- data.frame(time = 1:5, value = 5:9)
plot_line(df, x_str = "time", y_str = "value")
df2 <- data.frame(
  time = rep(1:5, 2),
  value = 1:10,
  group = rep(letters[1:2], each = 5)
)
plot_line(df2, x_str = "time", y_str = "value", color_str = "group")
plot_line(
  df2,
  x_str = "time", y_str = "value",
  color_str = "group", linetype_str = "group"
)

}
