% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair-plots.R
\name{plot_pairs}
\alias{plot_pairs}
\title{Custom ggpair plots}
\usage{
plot_pairs(
  data,
  columns,
  column_labels = NULL,
  color_lower = NULL,
  color_upper = NULL,
  color_label_lower = "",
  color_label_upper = "",
  color_scheme_lower = NULL,
  color_scheme_upper = NULL,
  title = "",
  point_size = 0.5,
  point_alpha = 0.5,
  cor_text_size = 3.5,
  subsample = 1,
  show_upper = TRUE,
  drop = TRUE
)
}
\arguments{
\item{data}{Data frame to use for plot.}

\item{columns}{Vector of column indicies or column names to plot.}

\item{column_labels}{(Optional) Label names to be displayed on strips.}

\item{color_lower}{(Optional) Data vector to use as colors for lower ggplot
panels.}

\item{color_upper}{(Optional) Data vector to use as colors for upper ggplot
panels.}

\item{color_label_lower}{Character string. Label for color legend title (used
in lower ggplot panels).}

\item{color_label_upper}{Character string for color_upper legend title (used
in upper ggplot panels).}

\item{color_scheme_lower}{(Optional) Vector of colors to set manual color
scheme corresponding to color_lower argument (i.e., the color scheme in the
lower panels). If \code{NULL} (default), viridis color scheme is used.}

\item{color_scheme_upper}{(Optional) Vector of colors to set manual color
scheme corresponding to color_upper argument (i.e., the color scheme in the
upper panels). If \code{NULL} (default), viridis color scheme is used.}

\item{title}{Character string. Title of plot.}

\item{point_size}{Point size for \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}.}

\item{point_alpha}{Alpha value for \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}.}

\item{cor_text_size}{Size of correlation text.}

\item{subsample}{Proportion of rows to subsample and plot.}

\item{show_upper}{Logical. Should we show subplots in upper panels?}

\item{drop}{Logical. Whether or not to drop factors with no observations.}
}
\value{
A \link[GGally:ggpairs]{GGally::ggpairs} object.
}
\description{
Creates pair plots, extending functionality beyond
\code{\link[GGally:ggpairs]{GGally::ggpairs()}} such as the ability to add two different color schemes
(one for the lower triangular subpanels and another for the upper
triangular subpanels).
}
\examples{
plot_pairs(data = iris, columns = 1:ncol(iris), color_lower = iris$Species)
plot_pairs(
  data = iris, columns = 1:4,
  color_lower = iris$Species, color_upper = rnorm(nrow(iris))
)

}
