% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-plots.R
\name{plot_point}
\alias{plot_point}
\title{Scatter plot wrapper for ggplot2.}
\usage{
plot_point(data, x_str, y_str, color_str = NULL, shape_str = NULL, ...)
}
\arguments{
\item{data}{Data frame to use for plot.}

\item{x_str}{Character string. Name of variable in \code{data} to plot on
x-axis.}

\item{y_str}{Character string. Name of variable in \code{data} to plot on
y-axis.}

\item{color_str}{Character string. Name of variable in \code{data}
to use as color aesthetic in plot.}

\item{shape_str}{Character string. Name of variable in \code{data}
to use as shape aesthetic in plot.}

\item{...}{Other arguments to pass to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}.}
}
\value{
A ggplot scatter plot.
}
\description{
Convenient wrapper around \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}.
}
\examples{
plot_point(iris, x_str = "Sepal.Width", y_str = "Sepal.Length")
plot_point(
  iris,
  x_str = "Sepal.Width",
  y_str = "Sepal.Length",
  color_str = "Species"
)
plot_point(
  iris,
  x_str = "Sepal.Width",
  y_str = "Sepal.Length",
  color_str = "Species",
  shape_str = "Species"
)

}
