% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-plots.R
\name{shared_plot_args}
\alias{shared_plot_args}
\title{Arguments that are shared by multiple wrapper functions}
\arguments{
\item{data}{Data frame to use for plot.}

\item{x_str}{Character string. Name of variable in \code{data} to plot on
x-axis.}

\item{y_str}{Character string. Name of variable in \code{data} to plot on
y-axis.}

\item{color_str}{Character string. Name of variable in \code{data}
to use as color aesthetic in plot.}

\item{fill_str}{Character string. Name of variable in \code{data}
to use as fill aesthetic in plot.}

\item{shape_str}{Character string. Name of variable in \code{data}
to use as shape aesthetic in plot.}

\item{linetype_str}{Character string. Name of variable in \code{data}
to use as linetype aesthetic in plot.}

\item{group_str}{Character string. Name of variable in \code{data}
to use as group aesthetic in plot.}

\item{fill}{Fill color. Only used if \code{fill_str} is \code{NULL}.}

\item{alpha}{Alpha value for transparency.}
}
\description{
Arguments that are shared by multiple wrapper functions
}
\keyword{internal}
